requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.17.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
attrs==24.2.0
Authlib==1.6.5
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.39.9
botocore==1.39.9
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
clang==19.1.7
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.70.0
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
et_xmlfile==2.0.0
expandvars==1.1.1
faiss-cpu==1.9.0.post1
fastapi==0.115.14
filelock==3.16.1
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.39.0
google-auth-oauthlib==1.2.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.0
grpcio==1.67.1
gspread==6.2.0
gunicorn==23.0.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-aws==0.2.31
langchain-community==0.3.18
langchain-core==0.3.75
langchain-huggingface==0.1.2
langchain-milvus==0.2.1
langchain-nvidia-ai-endpoints==0.3.17
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langfuse==2.58.4
langgraph==0.6.7
langgraph-checkpoint==2.1.2
langgraph-sdk==0.2.2
langsmith==0.3.45
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.7
mcp==1.13.1
mdurl==0.1.2
milvus-lite==2.5.1
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nvidia-nat==1.3.1
nvidia-nat-langchain==1.3.1
numpy==2.3.0
# NVIDIA CUDA packages - only install on Linux and Windows
nvidia-cublas-cu12==12.4.5.8; sys_platform == "linux" or sys_platform == "win32"
nvidia-cuda-cupti-cu12==12.4.127; sys_platform == "linux" or sys_platform == "win32"
nvidia-cuda-nvrtc-cu12==12.4.127; sys_platform == "linux" or sys_platform == "win32"
nvidia-cuda-runtime-cu12==12.4.127; sys_platform == "linux" or sys_platform == "win32"
nvidia-cudnn-cu12==9.1.0.70; sys_platform == "linux" or sys_platform == "win32"
nvidia-cufft-cu12==11.2.1.3; sys_platform == "linux" or sys_platform == "win32"
nvidia-curand-cu12==10.3.5.147; sys_platform == "linux" or sys_platform == "win32"
nvidia-cusolver-cu12==11.6.1.9; sys_platform == "linux" or sys_platform == "win32"
nvidia-cusparse-cu12==12.3.1.170; sys_platform == "linux" or sys_platform == "win32"
nvidia-cusparselt-cu12==0.6.2; sys_platform == "linux" or sys_platform == "win32"
nvidia-nccl-cu12==2.21.5; sys_platform == "linux" or sys_platform == "win32"
nvidia-nvjitlink-cu12==12.4.127; sys_platform == "linux" or sys_platform == "win32"
nvidia-nvtx-cu12==12.4.127; sys_platform == "linux" or sys_platform == "win32"
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.109.1
openinference-semantic-conventions==0.1.21
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
openpyxl==3.1.5
orjson==3.10.12
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pkce==1.0.3
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
prettytable==3.15.1
propcache==0.2.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.6.1
pydantic_core==2.33.2
Pygments==2.19.2
pymilvus==2.5.14
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
ragas==0.2.15
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.4.5
scikit-learn==1.7.2
scipy==1.15.0
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
sse-starlette==2.4.1
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0; sys_platform == "linux" or sys_platform == "win32"
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
